cbi: fix to work inside/along with other std directives
Currently the directive is not working inside/along with other std
directives (such as ngRepeat or with {{variable}}).
The initialization is done at compile or pre-link traversing the
children, but they are not yet linked if (if they are not static).

Change initialization to linking fase and change logic to not traverse
the children. Instead each children calls the Map at the moment of
initialization.

While at it, move global regex variables to local scope, as now they are
only needed there.

Signed-off-by: Adrian Panella <ianchi74@outlook.com>