luci2-ui-base: cbi: cbiOptionCtrl.formValue() ignore $scope argument
In order to allow using formValue() as $scope.$watch() watchExpression,
ignore the newValue argument if it is an Angular $scope instance.

Since watchExpression functions get passed the current scope as first
argument, the formValue() function would otherwise assume the scope
argument to be a new value to set and propagate it to .rawValue.

After this change we can conveniently use formValue for watching value changes
in subcontrollers and the following code:

  $scope.$watch(
	function(scope) { return cbiOptionCtrl.formValue() },
	callback
  )

can be rewritten as:

  $scope.$watch(cbiOptionCtrl.formValue, callback);

Signed-off-by: Jo-Philipp Wich <jo@mein.io>