luci2-ui-base: cbi: fix updating errors
Revert the use of templates with trasnsclusion, as it is not yet
consistent with the way $scopes are used, as option/section directives
need to access its parent's scop, which is not possible with a
transcluded scope.

Fix syntax error in NetworkList template

Readd cbiMap.read() function, to complete the save path.

Signed-off-by: Adrian Panella <ianchi74@outlook.com>