luci2-ui-base: rpc: resolve individual promises in batch call
Currently when l2rpc is in batch mode, individual rpc calls return
promises that immidiately resolve to the "request" instead of the
"result".
This is different to the behaviour when calling in non batch mode, where
returned promises resolve to the called rpc method's result.

Standarize to allways return the method's result.
When in batch mode the individual call will return a promise that will
resolve to only that call's result, and the "flush" call will keep
resolving to an array of all calls results.

Additionally, complete removal of non standard angular.deferrable() use.

Signed-off-by: Adrian Panella <ianchi74@outlook.com>