ui-core: UCI: propage validation state across levels
Angular template forms require "input" contros to be in the same
template as the "form" or "modelGroup".
The config/section/option directives split the inputs and the form in
different templates.
These helper directives will be used to provide the parent
form/modelGroup to the child component and re-enable standard template
form's functionality.

Place mark on the groups if invalid (now a '*', should be replaced by a
badge once the component makes it to Material2).
Disables submiting form if invalid.

Signed-off-by: Adrian Panella <ianchi74@outlook.com>